More progress tinkering with how Iron does his method. - Entities now spray blood particles. - Blood particles spatter on the ground similar to Iron's way. - The overlapping issue that I once had has been fixed.
