+++ src/main/resources/assets/bloodybits/particles/blood_spatter_particles.json
@@ -1,0 +1,5 @@
+{
+  "textures": [
+    "bloodybits:spatter_0"
+  ]
+}
