--- src/main/resources/assets/bloodybits/particles/blood_spray_particle.json
@@ -1,5 +1,0 @@
-{
-  "textures": [
-    "bloodybits:spray"
-  ]
-}
