+++ src/main/resources/assets/bloodybits/particles/blood_spray_particles.json
@@ -1,0 +1,5 @@
+{
+  "textures": [
+    "bloodybits:spray_test"
+  ]
+}
